% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ALGORITHM.nonthrGSA}
\alias{ALGORITHM.nonthrGSA}
\title{A one sentence description of what your function does}
\usage{
ALGORITHM.nonthrGSA(sets = GROUPSod$groups, genes = GENESod,
  Q.up = seq(0.75, 0.95, 0.05), Q.do = rev(seq(0.05, 0.25, 0.05)),
  sigBord.sets = 0.01, adj.sets = "unadjusted", Neff = NULL,
  groupTest = Fisher.segm.test, Tmatrix = TTod.np.all$statistic)
}
\arguments{
\item{inputParameter1}{A description of the input parameter \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

