% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ALGORITHM.SIGnonthrGSA}
\alias{ALGORITHM.SIGnonthrGSA}
\title{A one sentence description of what your function does}
\usage{
ALGORITHM.SIGnonthrGSA(sets = GROUPSdo$groups, genes = GENESdo,
  sigBord.genes = 0.05, Q.up.l = 10, Q.do.l = 10, sigBord.sets = 0.01,
  adj.sets = "unadjusted", Neff = NULL, groupTest = Fisher.SIGsegm.test,
  Tobject = PVAL.do)
}
\arguments{
\item{inputParameter1}{A description of the input parameter \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

