% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chipAnnot}
\alias{chipAnnot}
\title{A one sentence description of what your function does}
\usage{
chipAnnot(chip = "hgu133plus2", geneIDs = rownames(PVAL.MeHg$up.q.values))
}
\arguments{
\item{inputParameter1}{A description of the input parameter \code{inputParameter1}}

\item{inputParameter2}{A description of the input parameter \code{inputParameter2}}
}
\value{
output A description of the object the function outputs
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

